// src/sources/ficr/dto/ficr-competition.dto.ts
export type FicrCompetitionDto = {
  Year: number;
  TeamCode: number;
  Description: string;
  Data: string;
  ma_Data: string;
  DSC: string;
  ID: number;
  ShowID: number;
  ProgramVersion: string;
  pi_LunghezzaVasca: number;
  pi_NumeroCorsie: number;
  Place: string;
  ma_DataRiferimento: string;
  ma_Descrizione: string;
};