// src/sources/ficr/dto/fict-tempo.dto.ts
export type FicrTempoDto = {
  DescrGara: string;
  DescrCat: string;
  ba_Descrizione: string;
  Corsia: number;
  NumConc: number;
  Metri: number;
  Tempo: string;
  PuntiA: number;
  PuntiS: number;
  Punti: number;
  Squadra: number;
  Frazione: number;
  Stato: number;
  Batteria: number;
  Categoria: string;
  TipoTurno: number;
  TipoGara: number;
  Staffetta: boolean;
  Sigla: string;
  Pos: number;
};